DB 변수 타입 수정 & 쿼리문 어트리뷰트 명 수정
테이블과 dto 맞추기위함

+++ src/main/java/com/nts/reservation/product/dto/Product.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Product {
-	private long id;
-	private long displayInfoId;
+	private int id;
+	private int displayInfoId;
 	private String placeName;
 	private String content;
 	private String description;
